Last commit

+++ Practica3/code/timer.h
@@ -14,6 +14,6 @@ int timer0_num_int;
 /*--- declaracion de funciones visibles del módulo timer.c/timer.h ---*/
 void timer_init(void);
 void timer0_empezar(void);
-int timer0_leer(void);
+long timer0_leer(void);
 
 #endif /* _TIMER_H_ */
